[Refactor] #11 - 로직 hooks 화
